server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt>=5.0.0",
    "black>=25.9.0",
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.3",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.37.0",
]
